Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> > Patch Set 14:
> > 
> > following kubernetes best practices, rootfs should be immutable so this shouldn't be done this way.
> > Can you provide me the link of your entrypoint so I can take a look?
> > 
> > thanks!
> 
> Sure: https://github.com/onap/vid/blob/master/deliveries/src/main/scripts/localize.sh
> 
> Which will lead you to this -- https://github.com/onap/vid/tree/master/deliveries/src/main/scripts

So, If I understand well your scripts, you are:

- unarchiving the war file into tomcat
- changing the configuration files with variables given by helm chart

First, it's immutable so it's ok but not perfect ;)
And you know that ONAP is about perfection :P

Let me propose this:
- we unarchive at Docker container generation
- we change the configuration at preboot

Would you be OK with that? 
If VID is used elsewhere than Kubernetes and you need your scripts for this part, we can change a little bit the script in order to not launch them if a env variable is set (I would prefer not to have to do that but it's up to you).

If you're OK, I can create the new Docker and start working on changing the helm chart!

Patch-set: 14