Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> @Krzysztof Opasiak, The release template note has been presented by Documentation PTL and validated with PTL
> 
> I agree that we should have links towards previous Release Notes. We can add links towards previosu releases on the bottom of the current release note.

Yeah I remember reviewing this one but I've not realized that we are changing this from being cumulative to per release.

I'm kind of concern if people realize this and if they put enough work while doing release notes for Frankfurt. Previously release notes were considered cumulative (at least there was such sentence in the release notes template) which means that if for example in Dublin release you document that there is an issue with certain workaround and then you never mention it in El Alto or Frankfurt it means that the issue is still there. Now when we change this release notes from being cumulative to "single shot" we need to merge all previous releases and explicitly once again mention the same bug with the same workaround.

Patch-set: 2
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>