Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > 2.1.23 has not been merged into master (https://gerrit.onap.org/r/c/oom/+/108322) and last time we merged a new AAF everything went wild.
> > > So I expect:
> > > 100% pods + HC + e2e tests before merging on Master
> > > Then:
> > > 100% pods + HC + e2e tests before merging on Frankfurt
> > 
> > We've run 5 times this on master already and you +2 it
> > I also see other changes that do impact AAF clients(like templates rework) that got merged (both on Master AND Frankfurt) without the criteria above,
> 
> Let me comment on this as I'm the author of those changes. I'd like to ensure you that all of them were first merged to master. There wasn't even a single case when we merged sth first to F and then to master.

That's not my point, I'm ok with merging on master first but I think that the requested criteria  (100% on all tests which I 've never seen on any change btw) seems unreasonable, we've all seen intermittent failures on the gating and have to take informed decisions.

> 
> Additionally helm templates do not bring new or change the functionality of a running ONAP instance. They only modify the deployment process as it was with the certInitializer. To be more precise the templates work has just changed the way how helm chart looks like but didn't change the way they look like after being rendered by helm (just as if you put a pieces of C code into macros and then call those macros instead of copy-pasting the code itself).

Well, a change is a change, if we follow your logic, we wouldn't need to run gating of say e2e on template changes, but we still do it, that's because 0 risk is not possible. Similarly, gating is not a silver bullet, it helps but we need to jointly decide on reasonable criteria, by looking at the overall picture.
> 
> Finally those changes fixed a bug in offline installer because the team reported that in their package they have aaf_agent image in many different versions.
> 
> > so yes there was a regression once - this can happen - we always have the ability to revert
> 
> How does this comply with the "Don't break the build" and "Don't break the master" recommendations coming from the TSC? I believe that the more testing we can give to the patch before it is merge the better source code we will have.

And I'm not questionning that, but we can't honestly say that we cover 100% of all possible paths for each patch, so there is always a possibility that something will fail.

> 
> >, but I don't see why rules should be changed just for this, all changes are calculated risks at this point. Not having this in Frankfurt means a couple of tickets will stay open (portal-894 and another one with policy)
> 
> Well personally I believe in the Debian release philosophy:
> https://wiki.debian.org/ReleaseWhenReady

Patch-set: 1
Reviewer: Gerrit User 5 <5@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>