Update patch set 1
Patch Set 1:

Well,
First of all, I’m not saying that we’re not trusting this specific patch
But, AAF Authz (with SO, SDC, AAI, DMaaP MR and SDN-C) are in the critical path for any ONAP deployment and thus are treated differently than other components. 
Second, some (not all I agree) of e2e tests of latest gates (the one that may be problematic here) have been at 100% so I don’t see why we couldn’t manage it here. 
Last, certInit patches were on master when gates were not as good as now and we had to take « risks » and I wasn’t happy about that that.
My goal is not to prove anything and retain forever this patch so let’s do this if you’re OK: we try (on both branch in parallel) up to tomorrow 3PM UTC. If at this time no tests were 100% successful we’ll then merge. 
oom_redeploy

Patch-set: 1
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>