Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1: Code-Review-1
> > > > 
> > > > 2.1.23 has not been merged into master (https://gerrit.onap.org/r/c/oom/+/108322) and last time we merged a new AAF everything went wild.
> > > > So I expect:
> > > > 100% pods + HC + e2e tests before merging on Master
> > > > Then:
> > > > 100% pods + HC + e2e tests before merging on Frankfurt
> > > 
> > > We've run 5 times this on master already and you +2 it
> > > I also see other changes that do impact AAF clients(like templates rework) that got merged (both on Master AND Frankfurt) without the criteria above,
> > 
> > Let me comment on this as I'm the author of those changes. I'd like to ensure you that all of them were first merged to master. There wasn't even a single case when we merged sth first to F and then to master.
> 
> That's not my point, I'm ok with merging on master first but I think that the requested criteria  (100% on all tests which I 've never seen on any change btw) seems unreasonable, we've all seen intermittent failures on the gating and have to take informed decisions.

Well personally I love green color especially in the gating:D
I've seen it couple of times recently I'd be happy to show you but I'm getting 404 on all those gating results;)

Even through gating is imperfect (as every system in the world) I believe that it provides a precious feedback that we shouldn't ignore to keep our "grass" green as long as possible.

> 
> > 
> > Additionally helm templates do not bring new or change the functionality of a running ONAP instance. They only modify the deployment process as it was with the certInitializer. To be more precise the templates work has just changed the way how helm chart looks like but didn't change the way they look like after being rendered by helm (just as if you put a pieces of C code into macros and then call those macros instead of copy-pasting the code itself).
> 
> Well, a change is a change, if we follow your logic, we wouldn't need to run gating of say e2e on template changes, but we still do it, that's because 0 risk is not possible.

Well personally I'd be extremely happy to have sth like "kubernetes linter" that would simply allow us to compare the state before and after our template changes. It would really save us a lot of gating resources. Unfortunately I'm not aware of such tool.

> Similarly, gating is not a silver bullet, it helps but we need to jointly decide on reasonable criteria, by looking at the overall picture.

I fully agree with you and we quite often merge patches if we know that the error is "typical" like for example DCAE HC failure but it also often hides some issues that patches are introducing and we've quite a hard time debugging and trying to figure out what broke the master so we are trying to avoid this.

We may be a little tougher on merging patches but it's because several times we got comments from people that as soon as a release is out master stops working. We believe that it's a bad thing and as we have now a pretty stable source code we are trying to make sure that it stays in this condition as long as possible.
 
> > 
> > Finally those changes fixed a bug in offline installer because the team reported that in their package they have aaf_agent image in many different versions.
> > 
> > > so yes there was a regression once - this can happen - we always have the ability to revert
> > 
> > How does this comply with the "Don't break the build" and "Don't break the master" recommendations coming from the TSC? I believe that the more testing we can give to the patch before it is merge the better source code we will have.
> 
> And I'm not questioning that, but we can't honestly say that we cover 100% of all possible paths for each patch, so there is always a possibility that something will fail.

Oh definitely! You can never be sure but it's like with the "extreme sports". There is a better chance that you will break your leg when you are ski jumping than when you are playing chess right?;)

Just like @Sylvain mentioned in a comment above. Our intention is not to hold this patch forever but to merge it as soon as possible when we are confident that we would not break the master. It's nothing personal to you or to AAF project, we have several other patches in pretty much the same "pending state" (for example SO)

> 
> > 
> > >, but I don't see why rules should be changed just for this, all changes are calculated risks at this point. Not having this in Frankfurt means a couple of tickets will stay open (portal-894 and another one with policy)
> > 
> > Well personally I believe in the Debian release philosophy:
> > https://wiki.debian.org/ReleaseWhenReady

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>