Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> results still show previous run timestamps, strange

If I look at the pipeline now (6:25 UTC), it said that it has been initiated 13h ago (17:25 UTC)
corresponds to the dashboard I have (17:21)
On gerrit the timestamp of redeploy command is at 16:18 then the pipeline start is 17:44
I think it is consistent
Maybe it took time to refresh the dashboard (or was in cache) as it is sync from another regularly but not in real time.

regarding the results 
healthcheck is at 66% due to 1 error in core and medium but small & full are at 100%
the error corresponds to Basic DMAAP Message Router PubSub Health Check 
This test is run in core/small/medium and is OK in small
We already observed in some cases a possible race condition making only of the 3 up&running
as it is working in small and full is working, it is OK

regarding E2E tests
- png-registrate 2/3:  There are possible init timer issue even if we got a full success on the master 1st of June => not blocking
- clearwater => issue due to known race condtion in SDC , certification get stuck
- freeradius_nbi: original error, the csar distribution failed {'omfComponentID': 'clamp', 'timestamp': '1591117188310', 'url': '/sdc/v1/catalog/services/Freeradius/1.0/artifacts/service-Freeradius-csar.csar', 'status': 'DEPLOY_ERROR', 'errorReason': 'null'}, as the distribution failed at one point, the test stops.
but the basic_vm was fully OK, including the distribution stage.

Patch-set: 5
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>