Update patch set 5
Patch Set 5:

@Krzysztof, @Christophe, the patch seems to be fine (I took a look on gating3).
Can you look at it and gives me your point of view?

I'm not really happy by adding the Intermediate CA this way but as long as we don't update aaf "authz" by asking pkcs12 also for AAF-SMS (and I think it requires a new version of AAF) I believe we'll have to do that.

Patch-set: 5