Update patch set 33
Patch Set 33:

> Patch Set 33:
> 
> sorry to "vomit" logs message but web page rendering is not working today,
> we're on it but in the meantime I thought you would be interested to have the results

This is exactly what I was looking for Sylvain.
The issue is with the $RA_IAK that is created in EJBCA. This does not match the RA_IAK in cmpServers.json in the certService. I don't know if putting $RA_IAK in a json file renders it properly. If I hardcode on both sides the flow works perfectly. Not sure how to get around this.

Patch-set: 33
Reviewer: Gerrit User 4893 <4893@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>