Update patch set 1
Patch Set 1:

Note : this is actually a cherry pick of the following change:
https://gerrit.onap.org/r/c/oom/+/107449

Unfortunately, since that change is closed, git review rejects this change unless I remove the Change-ID footer - breaking the relationship link.  I tried several times to preserve the link, after a previous -1 due to this not being shown as a cherry pick, but to no avail.  It appears removing the Change-ID is the only option.

This fix is needed by the SDN-R team.  I'd appreciate your merging it at your earliest convenience.

Patch-set: 1