Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> @Jorge
> Thank you for your work, it looks awesome but... For G release we want to mandate (requirement is already submitted, waiting for TSC vote next week) all projects to use certInitializer and retrieve their certificates at runtime then you would not need those changes right?

Ah, ok, thanks, I was not aware of it.
I have a question though.  If we get the certificates at runtime, those certificates will end up in the java keystore (truststore) in the JVM within the container. Do we care about administering the passwords that the JVM's keystore/truststore passwords with external configuration?   If we do, this work may still be relevant in the context on setting the passwords for the keystore/truststore, otherwise we can just abandon the review.

Patch-set: 2