Update patch set 4
Patch Set 4:

Hello Kevin, thanks for the patch. 
We try to have logs also to STDOUT in order to simplify logs collection. 

Do you think it would be feasible to add them in this particular merge request?

Also, I know that « logback » allows part of configuration to be done via variables. 
Is it possible here? If yes it would be awesome to be able to set the log level via an environment variable (LOG_LEVEL for example with default to INFO).  
Doing that would help to simplify debugging when there are some issues.

Patch-set: 4