Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Base on G release requirements there should be no hardcoded certificates in both docker image and OOM repo and everyone should use certInitializer instead. Thus before we consider merging this I'd like to make sure that you acknowledge that and would be great to have a jira ticket tracing this

Hi Krzysztof, Thanks! This check-in is to fix the SOL003 Adapter secure communication with the ETSI Catalog Manager through MSB in the Frankfurt. Due to the last-minute MSB certificate change, the tested secured communication has been broken. This fix will sync up the MSB certificate for Frankfurt to make the communication between the SOL003 Adapter and ETSI Catalog Manager works again. We put the same fixes to both Frankfurt (part of the Frankfurt maintenance release) and Guilin (for now). We acknowledge that Guilin provides additional security requirements and we will conform to the changes. For the Guilin, this is the base and we will conform the Guilin security mechanism.

Patch-set: 1
Reviewer: Gerrit User 2085 <2085@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>