[COMMON] Allow to use ' in mariadb-init
' is one of characters that are placed in passwords by our default
password generation algorith. As ' is a special character in SQL
we need to escape it before concatenating it with SQL command.

Let's also add set -e to fail if any of databases has not been created

Issue-ID: OOM-2436
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: Ida8c75639eaa4049e3f874d30666f1fe4dc02e12
(cherry picked from commit c5dac87fa301247928211d5944be22ae14bcd534)