Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> 5gbulk is using DCAE and it seems we had a failure with DCAE discussion:
> 
> https://gating-results.onap.eu/results//109470-4/xtesting-smoke-usecases-robot/5gbulkpm/5gbulkpm/log.html#s1-t8
> 
> let's oom_redeploy to be sure

The root cause here is that the Cloudify Manager API begins to return 500 status codes to the API requests that the bootstrap container makes to deploy the various DCAE service components. In this case, the dcae-ves-collector needed by the 5gbulk PM was not deployed.

We have seen this problem occasionally in the past, but only in the gating environment.

Patch-set: 4