Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > 5gbulk is using DCAE and it seems we had a failure with DCAE discussion:
> > > 
> > > https://gating-results.onap.eu/results//109470-4/xtesting-smoke-usecases-robot/5gbulkpm/5gbulkpm/log.html#s1-t8
> > > 
> > > let's oom_redeploy to be sure
> > 
> > The root cause here is that the Cloudify Manager API begins to return 500 status codes to the API requests that the bootstrap container makes to deploy the various DCAE service components. In this case, the dcae-ves-collector needed by the 5gbulk PM was not deployed.
> > 
> > We have seen this problem occasionally in the past, but only in the gating environment.
> 
> hey Jack,
> I did took a look on recent deployments of our dailies from Orange and DT and that's true it seems we do not have such error.
> 
> This happens only on gating but more and more often.
> 
> The thing is our (Orange) dailies (Master and Frankfurt) and the gating are using the exact same code for deploying on the same OpenStack :S
> 
> the only difference is that VM and Kubernernetes are "fresh" on dailies whereas we first remove previous onap + /dockerdata-nfs cleaning + namespace removal on our gating.
> 
> Is there a way to debug more this because two third of our issues on healtcheck are coming from this?

I am trying to work out how we might capture more information.  If the problem occurred in my local test environment (or in the Wind River environment), I could exec onto the CM container and explore the various logs for clues.   There are a number of components running in the container (breaking it up into proper microservices is on the Cloudify roadmap but not in the near future) with a number of logs in different locations.  Would it be possible to get an archive of the /var/log directory tree from an instance where the problem has occurred?

Patch-set: 4