Update patch set 1
Patch Set 1:

We have to wait that Ofir merges the dcae-ds PR so that we can release them.
The current PR will NOT work until we have DCAE-ds images.

I raised this PR as i'm off the next 2 weeks, so that my team can easily complete the maintenance release.

Christophe Closset will follow that

Patch-set: 1