Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5: Code-Review-1
> 
> (1 comment)
> 
> It would be awesome if you could provide also some docummentation for this template explaining how you would like it to work as I'm not sure if you realize all the consequences of current code

I'm sure there are consequences that I don't fully realize, so any guidance you can provide will be much appreciated. Would you like more details in the commit message or where would that documentation go?

The main consequence that I anticipate, is that wherever certInitializer is used and this new feature is enabled - we will get an init container doing the work of importing certificates. I would love to have this done only once, but not sure how to do that (let me know if you have ideas). I did discuss this with Sebastien, and the conclusion is that this is still better than the way we manage certificates today by manually adding them to the truststore for the component that needs it. Is this the consequence you have in mind or are there others?

Please advise

Patch-set: 5