Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> (1 comment)
> 
> > Patch Set 5: Code-Review-1
> > 
> > (1 comment)
> > 
> > It would be awesome if you could provide also some docummentation for this template explaining how you would like it to work as I'm not sure if you realize all the consequences of current code
> 
> I'm sure there are consequences that I don't fully realize, so any guidance you can provide will be much appreciated. Would you like more details in the commit message or where would that documentation go?

We tend to document all the stuff in a comment above the template itself. You can check for example common secret template to see how we did this.

> 
> The main consequence that I anticipate, is that wherever certInitializer is used and this new feature is enabled - we will get an init container doing the work of importing certificates. I would love to have this done only once, but not sure how to do that (let me know if you have ideas). I did discuss this with Sebastien, and the conclusion is that this is still better than the way we manage certificates today by manually adding them to the truststore for the component that needs it. Is this the consequence you have in mind or are there others?

So by saying consequences I meant mostly how you intend to behave if AAF is enabled and how when it is disabled.
 
Just to make sure that we are on the same page: I've nothing against this patch apart from this small catch that I put in the comment. I just wanted to make sure that the way how I believe it works is exactly the thing that you expect from it in all conditions where this can be used;)

Patch-set: 5
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>