Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > Any reason to keep it as a sidecar instead of having it as a separated pod?
> > > BTW
> > > What's the role of this nginx? Couldn't it be replaced with ingress?
> > 
> > The main application here is an uwsgi application which is simply an app server. It needs a web server to run. Nginx is playing the role of that web server.
> 
> OK but then why not put nginx in a separate pod and let it speak to the service that just uses uwsgi protocol?
> 
> this way it is easier to debug as with sidecar container the pod is not failing as long as there is at least one sidecar running

Does this mean, the pod with uwsgi server is exposed via a kubernetes service and the nginx pod uses the service to communicate to the uswgi server?

Nginx also provides the SSL termination for the application. My concern is exposing uswgi app as a service directly. The uwsgi best practice suggests not to expose the socket speaking uswgi protocol to the external world. This channel allows for dynamic loading of applications which may lead to ACE attacks. 

I am not able to understand the second point. With multiple containers in a pod, if one of them fails, it will restart that container. Even with one container per pod, when the container fails, the pod will not fail. It will simply restart the container

Patch-set: 1