[GLOBAL] Add sleep 3 to avoid race condition
We are facing some random failures in our gating due to race condition
between our makefile and helm startup. To prevent that let's add a
sleep 3 after helm serve.

Yes we realize that it's not the proper solution but it's an easy win
which is very readable compared to placing a proper solution in the makefile.

Issue-ID: OOM-2511
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: If6b3332ea9f51e997f2a6057fff9d481b425104a