Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Based on https://gating-results.onap.eu/results//110437-2/infrastructure-healthcheck/k8s/pod-onap-sdnc-0-sdnc-logs.html I would say it is working now.

Yes I agree, cool! thanks

> 
> I noticed a few plugins have quite noisy startup (printing whole stack track during warning). If I find time I would clean that up.

OK, nice.

> 
> The console appender is using the pattern defined at
> https://gerrit.onap.org/r/gitweb?p=oom.git;a=blob;f=kubernetes/sdnc/resources/config/log/org.ops4j.pax.logging.cfg;hb=HEAD#l50
> 
> We could attempt switching to JSON or XML if desired.

we can try on this one yes, I've seen the pattern but wasn't sure on how to change it

> 
> One thing to consider is right now each different type of log has its own pattern. By type of log I mean audit vs metric vs error.
> 
> Right now metric and audit would not be getting printed on the console (stdout) at all right now. Right now those are not going to the root logger and the console appender is only getting log statements from the root logger.

So If I understand well, "debug.log", "error.log", "audit.log" have a different pattern in the file (as we can see in org.ops4j.pax.logging.cfg)
And on STDOUT, we'll see only logs with INFO and up (per default) with another specific pattern. And only if they are sent to the root logger (which is not the case for audit and metric).

Is it right?

I believe we can stick with that for now and work on the "best" way to generate the logs.
So let's try JSON or XML (preference to JSON personnally) and see what we have!

thanks again

Patch-set: 2