Update patch set 2
Patch Set 2:

we can try on this one yes, I've seen the pattern but wasn't sure on how to change it

Updating the pattern is very easy, patterns are only relevant if using a PatternLayout. I used one which was already in the file, but if we want something custom I can change it.

If we switch to JSON or XML that is a different kind of layout. It comes with some default formatting. You can provide custom formatting, but it isn't through patterns, but a custom formatter class.

So If I understand well, "debug.log", "error.log", "audit.log" have a different pattern in the file (as we can see in org.ops4j.pax.logging.cfg)
Correct.

And on STDOUT, we'll see only logs with INFO and up (per default) with another specific pattern. 
Not sure if saying "and up" is the right way. Trace has the integer logging level of 600.
Error has the integer logging level of 200.
If logging level is set to INFO you will get all logs for INFO and anything more severe (lower logging level).

So with INFO you get INFO, WARN, ERROR, FATAL.

https://logging.apache.org/log4j/2.x/manual/customloglevels.html for greater detail.

And only if they are sent to the root logger (which is not the case for audit and metric).
Correct. 

I can check the loggers for audit and metric to start sending records to the ConsoleAppender if desired.

I believe we could define multiple console appenders each with a different pattern layout.

Patch-set: 2
Reviewer: Gerrit User 1219 <1219@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>