Update patch set 2
Patch Set 2: Code-Review-2

This patch does not change anything in terms how environment variables are processed inside the container.

The motivation behind the requirement that you are trying to cover is to make sure that the actual config files that are used by your application are provided as configmaps not just the values.

Just to be clear: technically I'm not against this patch I'm happy to merge it but I wanted to make sure that you realize that it's not enough to cover this requirement

Patch-set: 2
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-2