Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Code-Review-1
> > > 
> > > SDNC Healtcheck is KO. Don't we have to change it also?
> > 
> > I think I disabled that healthcheck by removing the file kubernetes/consul/resources/config/consul-agent-config/sdnc-portal-health.json ... at least, it seemed to me like that file was driving that health check?
> 
> Also, we disabled that pod in El Alto - it was back briefly in Frankfurt, but we ended up disabling it again.  So I think we should be okay

Sorry, I should clarify ....

The SDNC healthcheck that is failing is not related to the admin portal (sorry, I was looking at the
infrastructure health checks, saw they passed, and thought you were concerned that we also needed to remove the check for the admin portal from there).  That failure is almost certainly a timing issue -the healthcheck was probably sent before the SDNC OpenDaylight container was fully up.  So, the call to the SLI-API:healthcheck endpoint isn't returning a JSON payload yet.  It would be best if that healthcheck either retried a few times or waited for a few minutes (5 mins should do - that's probably even overkill).  I would not expect that issue to be related to this change, since we had already disabled starting up the portal (so it's not likely that the timing would be different based on this change) - but if there is something I missed that would help, I'm certainly happy to amend this change to address it.

Patch-set: 1