[MariaDB] Create NFVO Database Within MariaDB
Adding an additional SH script in order to create the NFVO database, which will be used by the SO ETSI NFVO NS LCM.
It will be used to store required information on any NFVOs.

The creation of the NFVO Database should not have an effect on other databases, as their creations have not been modified.

Also renamed the Create so_user and so_admin SH scripts to start with 98, 99 in order to trigger after any new scripts are added in the future.

Issue-ID: SO-3159
Change-Id: Idbf09db294d5bbc9e7dca063bc08040ce2bc8568
Signed-off-by: rope252 <gareth.roper@est.tech>