[SO-ETSI-NFVO] Add SO DB and NFVO DB Credentials to NFVO NS LCM Pod for DB Communication
All of the credentials added have been accessed using the respective secrets. (dev-so-db-user-creds etc.)
In order to avoid hardcoding passwords etc.

I have added 2 jdbcUrl's for accessing the NFVO DB and the CamundaBPMN DB. These will be needed for code going into the SO codebase.

This change should not have an effect on any other charts, as properties have only been added to ETSI NFVO Charts.

Issue-ID: SO-3159
Change-Id: Ic6e0002ce64b872ac0c44fa1f124304c392e6a8b
Signed-off-by: rope252 <gareth.roper@est.tech>