Update patch set 8
Patch Set 8:

The certificates are getting generated in the init container in opt/app/osaaf, but the mount point is a different directory. Looks like configured mountPath in the values.yaml is being updated in both init-container and the application container. mountPath value should only be applied to application container.

Patch-set: 8