Update patch set 8
Patch Set 8:

Hi Sylvain,

Currently the mountPath field is of no use. If we modify it, the correct directory will not get mounted to the app container. Seems like a bug to me.

There is a minor change reqired in cert-init chart. Can I create another patch to address it?

Patch-set: 8