Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> it seems OK for me except vid seems to fail...
> @Ikram, can you take a look?

Just took a deeper look on this and found the root cause.

VID follows this awful patter of using sed in docker entrypoint to transfer the settings from envs to config files.

With this patch the master password string has changed which influenced all automatically generated passwords and now the VID DB password contains special characters that are interpreted by sed an causes destruction in vid config file.

As all those scripts are run from inside a docker container I've filled a jira ticket for VID team to fix this:
https://jira.onap.org/browse/VID-890

Patch-set: 1