Update patch set 9
Patch Set 9:

Hi,

Need a little suggestion. The key file seems to have 400 permission with user root and group root. Is it advisable to change the permission to 444 using add_aaf_config script, so that the container user can read it. Or, Can I create an init container to give read permission to the user?

Patch-set: 9