[SDNC] Fixes to aaiclient.properties
This change fixes the issue seen where SDNC is trying to reach
AAI via aai.api.simpledemo.onap.org endpoint.  The root cause was
that the mount for the aaiclient.properties file was incorrect
(ccsdkConfigDir instead of configDir), causing that version to
be ignored in favor of the version bundled into the docker image.

Issue-ID: SDNC-1354
Signed-off-by: Dan Timoney <dtimoney@att.com>
Change-Id: Ib9f9f51fd36150f2fe82323a767b28703a1a9390