[SDNC] Temporary disable sdnrdb
As diagnosed today if sdnc is enabled a different entrypoint for SDNC
is used. Unfortunately this makes SDNC reach AAI using a bad URL which
causes our e2e tests to fail.

As a temporary solution let's disable sdnrdb and hope that the team
will be able to debug this further and reenable it.

Issue-ID: SO-3216
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>
Change-Id: Ib2cda39aefea84a4c930d996e53429f3f64a86ef